--- a/client.go
+++ b/client.go
@@ -16,20 +16,14 @@ import (
 const (
 	defaultTimeout = 3000 // three seconds
 
-	// waiter "high water mark,"
-	// maximum number of queued writes.
-	// we start throttling requests after
-	// we reach this point.
+	// waiter "high water mark"
 	// TODO(maybe): make this adjustable.
 	waiterHWM = 32
 )
 
-type clientState byte
-
 const (
-	zeroState clientState = iota
-	closedState
-	openState
+	clientClosed = iota
+	clientOpen
 )
 
 var (
@@ -47,18 +41,6 @@ var (
 	// size is larger than 65,535 bytes.
 	ErrTooLarge = errors.New("synapse: message body too large")
 )
-
-// AsyncResponse is returned by
-// calls to (*Client).Async
-type AsyncResponse interface {
-	// Read reads the response to the
-	// request into the object, returning
-	// any errors encountered. Read blocks
-	// until a response is received. Calling
-	// Read more than once will cause undefined behavior.
-	// Calling Read(nil) discards the response.
-	Read(out msgp.Unmarshaler) error
-}
 
 // Dial creates a new client by dialing
 // the provided network and remote address.
@@ -91,10 +73,9 @@ func DialTLS(network, raddr string, timeout int64, config *tls.Config) (*Client,
 func NewClient(c net.Conn, timeout int64) (*Client, error) {
 	cl := &Client{
 		conn:    c,
-		pending: make(map[uint64]*waiter, waiterHWM),
 		writing: make(chan *waiter, waiterHWM),
 		done:    make(chan struct{}),
-		state:   openState,
+		state:   clientOpen,
 	}
 	go cl.readLoop()
 	go cl.writeLoop()
@@ -119,50 +100,41 @@ func NewClient(c net.Conn, timeout int64) (*Client, error) {
 // Client is a client to
 // a single synapse server.
 type Client struct {
-	conn    net.Conn           // connection
-	csn     uint64             // sequence number; atomic
-	mlock   sync.Mutex         // to protect 'mlock' and 'state'
-	pending map[uint64]*waiter // map seq number to waiting handler
-	writing chan *waiter       // queue to write to conn; size is effectively HWM
-	done    chan struct{}      // closed during (*Client).Close to shut down timeoutloop
-	state   clientState        // open, closed, etc.
+	conn    net.Conn       // connection
+	csn     uint64         // sequence number; atomic
+	writing chan *waiter   // queue to write to conn; size is effectively HWM
+	done    chan struct{}  // closed during (*Client).Close to shut down timeoutloop
+	wg      sync.WaitGroup // outstanding client procs
+	state   uint32         // open, closed, etc.
+	pending wMap           // map seq number to waiting handler
 }
 
 // used to transfer control
 // flow to blocking goroutines
 type waiter struct {
-	next   *waiter    // next in linked list, or self
+	next   *waiter    // next in linked list, or nil
 	parent *Client    // parent *client
+	seq    uint64     // sequence number
 	done   sync.Mutex // for notifying response; locked is default
 	err    error      // response error on wakeup, if applicable
 	in     []byte     // response body
-	reap   bool       // can reap for timeout (see: (*client).timeoutLoop())
-	_      [sizeofPtr - 1]byte
+	reap   bool       // can reap for timeout
+	static bool       // is part of the statically allocated arena
 }
 
 // Close idempotently closes the
 // client's connection to the server.
-// Goroutines blocked waiting for
-// responses from the server will be
-// unblocked with an error.
+// Close returns once every waiting
+// goroutine has either received a
+// response or timed out; goroutines
+// with requests in progress are not interrupted.
 func (c *Client) Close() error {
-	// don't let blocking reads/writes
-	// prevent another goroutine from
-	// unlocking the map lock (eventually)
-	c.conn.SetWriteDeadline(time.Now().Add(1 * time.Second))
-	c.conn.SetReadDeadline(time.Now().Add(1 * time.Second))
-
-	c.mlock.Lock()
-	if c.state == closedState {
-		c.mlock.Unlock()
+	// there can only be one winner of the Close() race
+	if !atomic.CompareAndSwapUint32(&c.state, clientOpen, clientClosed) {
 		return ErrClosed
 	}
-	c.state = closedState
-	for _, val := range c.pending {
-		val.err = ErrClosed
-		val.done.Unlock()
-	}
-	c.mlock.Unlock()
+
+	c.wg.Wait()
 	close(c.done)
 	close(c.writing)
 	return c.conn.Close()
@@ -170,21 +142,22 @@ func (c *Client) Close() error {
 
 // close with error
 // sets the status of every waiting
-// goroutine to 'err' and unblocks it
+// goroutine to 'err' and unblocks it.
 func (c *Client) closeError(err error) {
-	err = fmt.Errorf("synapse: fatal error: %s", err)
-
-	c.mlock.Lock()
-	if c.state == closedState {
-		c.mlock.Unlock()
+	if !atomic.CompareAndSwapUint32(&c.state, clientOpen, clientClosed) {
 		return
 	}
-	c.state = closedState
-	for _, val := range c.pending {
-		val.err = err
-		val.done.Unlock()
+
+	err = fmt.Errorf("synapse: fatal error: %s", err)
+
+	// we can't actually guarantee that we will preempt
+	// every goroutine, but we can try.
+	c.pending.flush(err)
+	for c.pending.length() > 0 {
+		c.pending.flush(err)
 	}
-	c.mlock.Unlock()
+
+	c.wg.Wait()
 	close(c.done)
 	close(c.writing)
 	c.conn.Close()
@@ -232,18 +205,8 @@ func (c *Client) readLoop() {
 			continue
 		}
 
-		// note: this lock
-		// is also acquired by
-		// writing goroutines, which
-		// may block while holding the lock.
-		c.mlock.Lock()
-		w, ok := c.pending[seq]
-		if ok {
-			delete(c.pending, seq)
-		}
-		c.mlock.Unlock()
-
-		if !ok {
+		w := c.pending.remove(seq)
+		if w == nil {
 			if !c.do(bwr.Skip(sz)) {
 				return
 			}
@@ -329,17 +292,7 @@ func (c *Client) timeoutLoop(msec int64) {
 		case <-c.done:
 			return
 		case <-tick:
-			c.mlock.Lock()
-			for seq, w := range c.pending {
-				if w.reap {
-					delete(c.pending, seq)
-					w.err = ErrTimeout
-					w.done.Unlock()
-				} else {
-					w.reap = true
-				}
-			}
-			c.mlock.Unlock()
+			c.pending.reap()
 		}
 	}
 }
@@ -347,12 +300,16 @@ func (c *Client) timeoutLoop(msec int64) {
 // write a command to the connection - works
 // similarly to standard write()
 func (w *waiter) writeCommand(cmd command, msg []byte) error {
+	w.parent.wg.Add(1)
+	if atomic.LoadUint32(&w.parent.state) == clientClosed {
+		return ErrClosed
+	}
+	seqn := atomic.AddUint64(&w.parent.csn, 1)
+
 	cmdlen := len(msg) + 1
 	if cmdlen > maxMessageSize {
 		return ErrTooLarge
 	}
-
-	seqn := atomic.AddUint64(&w.parent.csn, 1)
 
 	// write frame + message
 	need := leadSize + cmdlen
@@ -364,26 +321,23 @@ func (w *waiter) writeCommand(cmd command, msg []byte) error {
 	putFrame(w.in, seqn, fCMD, cmdlen)
 	w.in[leadSize] = byte(cmd)
 	copy(w.in[leadSize+1:], msg)
-	w.reap = false
 
-	// the whole write process must
-	// be atomic with respect to
-	// the map write and channel send
-	// in order to prevent racing on
-	// (*Client).Close
+	w.reap = false
+	w.seq = seqn
+
 	p := w.parent
-	p.mlock.Lock()
-	if p.state != openState {
-		p.mlock.Unlock()
-		return ErrClosed
-	}
-	p.pending[seqn] = w
+	p.pending.insert(w)
 	p.writing <- w
-	p.mlock.Unlock()
 	return nil
 }
 
 func (w *waiter) write(method string, in msgp.Marshaler) error {
+	w.parent.wg.Add(1)
+	if atomic.LoadUint32(&w.parent.state) == clientClosed {
+		return ErrClosed
+	}
+	sn := atomic.AddUint64(&w.parent.csn, 1)
+
 	var err error
 
 	// save bytes up front
@@ -412,33 +366,13 @@ func (w *waiter) write(method string, in msgp.Marshaler) error {
 		return ErrTooLarge
 	}
 
-	sn := atomic.AddUint64(&w.parent.csn, 1)
 	putFrame(w.in, sn, fREQ, olen)
 	w.reap = false
+	w.seq = sn
 
-	// the whole write process must
-	// be atomic with respect to
-	// the map write and channel send
-	// in order to prevent racing on
-	// (*Client).Close
-	//
-	// it's possible that writes will
-	// block here for a while, but this
-	// is actually to our advantage: there
-	// exists a high-water mark after which
-	// new requests will be bottlenecked by
-	// locking around map access, which means
-	// that we cannot increase memory footprint
-	// without bound.
 	p := w.parent
-	p.mlock.Lock()
-	if p.state != openState {
-		w.parent.mlock.Unlock()
-		return ErrClosed
-	}
-	p.pending[sn] = w
+	p.pending.insert(w)
 	p.writing <- w
-	p.mlock.Unlock()
 	return nil
 }
 
@@ -459,10 +393,12 @@ func (w *waiter) read(out msgp.Unmarshaler) error {
 func (w *waiter) call(method string, in msgp.Marshaler, out msgp.Unmarshaler) error {
 	err := w.write(method, in)
 	if err != nil {
+		w.parent.wg.Done()
 		return err
 	}
 	// wait for response
 	w.done.Lock()
+	w.parent.wg.Done()
 	if w.err != nil {
 		return w.err
 	}
@@ -479,17 +415,25 @@ func (c *Client) Call(method string, in msgp.Marshaler, out msgp.Unmarshaler) er
 	return err
 }
 
+// errors specific to commands
+var (
+	errNoCmd      = errors.New("synapse: no response CMD code")
+	errInvalidCmd = errors.New("synapse: invalid command")
+	errUnknownCmd = errors.New("synapse: unknown command")
+)
+
 // doCommand executes a command from the client
 func (c *Client) sendCommand(cmd command, msg []byte) error {
 	w := waiters.pop(c)
 	err := w.writeCommand(cmd, msg)
 	if err != nil {
+		c.wg.Done()
 		return err
 	}
 
 	// wait
 	w.done.Lock()
-
+	c.wg.Done()
 	if w.err != nil {
 		return w.err
 	}
@@ -497,20 +441,20 @@ func (c *Client) sendCommand(cmd command, msg []byte) error {
 	// bad response
 	if len(w.in) == 0 {
 		waiters.push(w)
-		return errors.New("no response CMD code")
+		return errNoCmd
 	}
 
 	ret := command(w.in[0])
 
 	if ret == cmdInvalid || ret >= _maxcommand {
 		waiters.push(w)
-		return errors.New("command invalid")
+		return errInvalidCmd
 	}
 
 	act := cmdDirectory[ret]
 	if act == nil {
 		waiters.push(w)
-		return errors.New("unknown CMD code returned")
+		return errUnknownCmd
 	}
 
 	act.Client(c, w.in[1:])
@@ -523,31 +467,4 @@ func (c *Client) sendCommand(cmd command, msg []byte) error {
 // didn't respond appropriately
 func (c *Client) ping() error {
 	return c.sendCommand(cmdPing, nil)
-}
-
-// AsyncResponse.Read implementation
-func (w *waiter) Read(out msgp.Unmarshaler) error {
-	w.done.Lock()
-	if w.err != nil {
-		return w.err
-	}
-	err := w.read(out)
-	waiters.push(w)
-	return err
-}
-
-// Async sends a request to the server, but does not
-// wait for a response. The returned AsyncResponse object
-// should be used to decode the response. After the first
-// call to Read(), the returned AsyncResponse object becomes
-// invalid. Async is safe to call from multiple goroutines
-// simultaneously.
-func (c *Client) Async(method string, in msgp.Marshaler) (AsyncResponse, error) {
-	w := waiters.pop(c)
-	err := w.write(method, in)
-	if err != nil {
-		waiters.push(w)
-		return nil, err
-	}
-	return w, nil
 }