Before, pending requests live in a `map[uint64]*waiter`. While that was expedient, it was not fast.

The implementation of sequence numbers on the client side gives us the opportunity to take advantage of the fact that sequence numbers are monotonic and thus uniformly distributed.

What I have here is a static array of thread-safe queues. The way it works is it takes the last `rootBits` bits (6, in this case) of the sequence number as an offset into the "root array." From there, it's just naive search through a queue (while holding the queue's lock). This has a couple notable properties:

 - Locking is per-list, so a 64-node map can have 64 concurrent mutators.
 - Waiters are already allocated, so the data structure itself does zero allocation *ever*. However, the structure itself is quite large.
 - O(1) insertion all the time.
 - O(n) average-case removal *technically*, but best-case (O(1)) is extremely common, because we insert at the tail and search for removal beginning at the head. (Insertion/removal pattern is approximately FIFO.)
 - The timeout "reap" process doesn't have to lock the entire map at once, so it's much friendlier to calls actually making forward progress. More broadly, most things that have to iterate over the map don't need to observe it in a perfectly consistent state, and so they don't cause as much contention.

So now for the good stuff:

`GOMAXPROCS=4`

| Benchmark | Before | After | Delta |
|:---------------:|:---------:|:------:|:-------:|
| TCP Echo | 3005 ns | 1677 ns | -44% |
| Unix no-op | 1370 ns | 1190 ns | -13% |
| Pipe no-op | 1228 ns | 899 ns | -27% |

`GOMAXPROCS=1`

| Benchmark | Before | After | Delta |
|:---------------:|:---------:|:------:|:-------:|
| TCP Echo | 3866 ns | 3788 ns | -2% |
| Unix no-op | 2231 ns | 2071 ns | -7% |
| Pipe no-op | 1646 ns | 1569 ns | -5% |

So it *is* faster in the un-contended case as well.

Warning: This is still very much a work in progress. Gross code ahead.

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/tinylib/synapse/20)
<!-- Reviewable:end -->