@philhofer 2015-01-29
Now that map access is atomic, I think I can find a way to set up client state to be managed via atomics rather than with locking, which might buy a little more perf as well.