@philhofer 2015-01-30
Insertion is now done at the tail of each bucket's queue, and removal search begins at the top. This means that, for any *precisely* FIFO usage pattern, both insertion and removal are constant-time. Even with some stochasticity, constant-time removal is extremely common.

I've also added some benchmarks that compare this map implementation to the standard map, and it runs about 2.5x faster on my computer.