@philhofer 2015-02-06
@ttacon I have temporarily removed `Async` from the client implementation, for two reasons:

 1. Particularly with the new map implementation, it allows poorly-written client code to do horrible racy things, because the object that implements `AsyncResponse` is recycled.
 2. Clients can pretty easily implement this functionality themselves with goroutines and waitgroups. (And, all else being equal, less exported stuff is better, right?)

Let me know if you have strong objections.