@ttacon 2015-02-06
Naw, that sounds entirely reasonable to me. Giving a user fewer ways to write unfortunate code is always better, especially when it doesn't impact a library's usability.